Add abstract support
Adds support for `x-abstract` on `ObjectSchema` and `AllOfSchema`.

Allows a developer to declare that a schema is considered abstract. This could be useful for code generation when a developer wants to generate an interface or a protocol instead of a concrete class from a schema definition.

`x-abstract` in combination with `discriminator` allows a developer to declare interesting patterns, like arrays of mixed objects that share an interface and a discrminator to determine (which determines what type of object it should be parsed as).